Massive GroupModes refactor
The old GroupMode classes have each been turned into either a SplitMode or an AssignMode replicating their original functioning.

GroupMode is now a class holding a list of SplitModes followed by an AssignMode.

Also added type annotations to some of the classes/methods in groupmodes.py, to make it easier to make sense of it all.